changed data type to remove warning

+++ RRiOS7BackportFoundation/RRiOS7BackportFoundation/NSData.m
@@ -97,7 +97,7 @@
     unsigned char *outputBytes = (unsigned char *)[outputData mutableBytes];
     
     int accumulator = 0;
-    long long outputLength = 0;
+    NSUInteger outputLength = 0;
     unsigned char accumulated[] = {0, 0, 0, 0};
     for ( long long i = 0; i < inputLength; i++ ) {
         unsigned char decoded = lookup[inputBytes[i] & 0x7F];
